AAE-24331 TaskSearch ITs in dedicated test class, boolean variable test cases
